move api interaction to client module
